Se agrego la posibilidad de

+++ argy_reviews/api/views.py
@@ -1,11 +1,13 @@
 from .models import Post, Review, Report
 from .serializers import PostSerializer, ReviewSerializer, UserSerializer
 
-from rest_framework import status
 from rest_framework.decorators import api_view, permission_classes, authentication_classes
+from rest_framework.permissions import AllowAny, IsAuthenticated
+
+from rest_framework import status
 from rest_framework.response import Response
 from rest_framework import generics
-from rest_framework.permissions import AllowAny, IsAuthenticated
+
 from rest_framework.authtoken.models import Token
 from rest_framework.authentication import TokenAuthentication
 
@@ -16,6 +18,31 @@ from django.utils import timezone
 from datetime import timedelta
 from django.db.models import Avg, Max
 
+# TODO
+# - [x] Create a model for the Report
+
+
+# UserList 
+class UserList(generics.ListCreateAPIView):
+    queryset = User.objects.all()
+    serializer_class = UserSerializer
+
+# User-Detail
+@api_view(['GET'])
+def user_detail(request, user_pk):
+    """
+    Retrieve a user instance.
+    """
+    if request.method == 'GET':
+        try:
+            user = User.objects.get(pk=user_pk)
+        except User.DoesNotExist:
+            return Response(status=status.HTTP_404_NOT_FOUND)
+
+        serializer = UserSerializer(user)
+        return Response(serializer.data)
+
+# Carousels
 @api_view(['GET'])
 @permission_classes([AllowAny])
 def get_carousels_data(request):
@@ -52,22 +79,10 @@ def get_carousels_data(request):
     # Devuelve los datos como una respuesta JSON
     return Response(data)
 
-@api_view(['GET'])
-def user_detail(request, user_pk):
-    """
-    Retrieve a user instance.
-    """
-    if request.method == 'GET':
-        try:
-            user = User.objects.get(pk=user_pk)
-        except User.DoesNotExist:
-            return Response(status=status.HTTP_404_NOT_FOUND)
-
-        serializer = UserSerializer(user)
-        return Response(serializer.data)
-
+# Post-List
 @api_view(['GET', 'POST'])
-@permission_classes([AllowAny])
+@authentication_classes([TokenAuthentication])
+@permission_classes([IsAuthenticated])
 def post_list(request):
     """
     List all posts or create a new post.
@@ -82,9 +97,11 @@ def post_list(request):
         if serializer.is_valid():
             serializer.save(owner=request.user)
             return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)  
 
+# Post-Detail   
 @api_view(['GET', 'PUT', 'DELETE'])
+@permission_classes([AllowAny])
 def post_detail(request, post_pk):
     """
     Retrieve, update or delete a post.
@@ -95,20 +112,31 @@ def post_detail(request, post_pk):
         return Response(status=status.HTTP_404_NOT_FOUND)
 
     if request.method == 'GET':
+        # Allows acceses to lecture without authentication
         serializer = PostSerializer(post)
         return Response(serializer.data)
 
-    elif request.method == 'PUT':
-        serializer = PostSerializer(post, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    elif request.method in ['PUT', 'DELETE']:
+        # Requerires atutentication to update and delete a post
+        if not request.user.is_authenticated:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+        
+        # Verifies if the user is the owner of the post
+        if post.owner != request.user:
+            return Response(status=status.HTTP_403_FORBIDDEN)
 
-    elif request.method == 'DELETE':
-        post.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if request.method == 'PUT':
+            serializer = PostSerializer(post, data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
+        elif request.method == 'DELETE':
+            post.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+
+# Reviews-List
 @api_view(['GET', 'POST'])
 @permission_classes([AllowAny])
 def reviews_list(request, post_pk):
@@ -132,6 +160,7 @@ def reviews_list(request, post_pk):
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
+# Review-Detail
 @api_view(['GET', 'PUT', 'DELETE'])
 def review_detail(request, post_pk, review_pk):
     """
@@ -156,14 +185,14 @@ def review_detail(request, post_pk, review_pk):
     elif request.method == 'DELETE':
         review.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
-    
+
+# Login
 @api_view(['POST'])
 @permission_classes([AllowAny])
 def login(request):
     """
     Login a user.
     """
-    
     user = get_object_or_404(User, username=request.data['username'])
     
     if not user.check_password(request.data['password']):
@@ -172,8 +201,9 @@ def login(request):
     token, created = Token.objects.get_or_create(user=user)
     serializer = UserSerializer(instance=user)
     
-    return Response({'token': token.key, "user": serializer.data}, status=status.HTTP_200_OK)
-    
+    return Response({"user": serializer.data, "token": token.key}, status=status.HTTP_200_OK)
+
+# Register
 @api_view(['POST'])
 @permission_classes([AllowAny])
 def register(request):
@@ -193,6 +223,7 @@ def register(request):
         return Response({'token': token.key, "user": serializer.data}, status=status.HTTP_201_CREATED)
     return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
+# Profile
 @api_view(['GET'])
 @authentication_classes([TokenAuthentication])
 @permission_classes([IsAuthenticated])
@@ -203,10 +234,6 @@ def profile(request):
     serializer = UserSerializer(instance=request.user)
     
     return Response(serializer.data, status=status.HTTP_200_OK)
-    
-class UserList(generics.ListCreateAPIView):
-    queryset = User.objects.all()
-    serializer_class = UserSerializer
     
 # @api_view(['GET', 'POST'])
 # @permission_classes([AllowAny])
